Changed: Deduplicated some code
