+++ Persona.Merger.Common/Patching/Tbl/FieldResolvers/P3P/P3PTblPatcher.cs
@@ -13,6 +13,7 @@ using Persona.Merger.Patching.Tbl.FieldResolvers.P3P.Effect;
 using Persona.Merger.Patching.Tbl.FieldResolvers.P3P.AICalc;
 using Persona.Merger.Patching.Tbl.FieldResolvers.P3P.Item;
 using Persona.Merger.Patching.Tbl.FieldResolvers.P3P.Message;
+using static Persona.Merger.Patching.Tbl.FieldResolvers.TblPatcherCommon;
 
 namespace Persona.Merger.Patching.Tbl.FieldResolvers.P3P;
 
@@ -161,13 +162,7 @@ public struct P3PTblPatcher
                 P3PTblSegmentFinder.Populate(tblData, segmentCount, originalSegments);
 
             // Convert original segments into Memory<T>.
-            var segments = new Memory<byte>[segmentCount];
-            for (int x = 0; x < segmentCount; x++)
-            {
-                ref var originalSegment = ref originalSegments[x];
-                var offset = originalSegment.Pointer - tblData;
-                segments[x] = new Memory<byte>(TblData, (int)offset, originalSegment.Length);
-            }
+            var segments = ConvertSegmentsToMemory(segmentCount, originalSegments, tblData, TblData);
 
             // Apply Patch(es).
             for (int x = 0; x < segmentCount; x++)
@@ -180,17 +175,10 @@ public struct P3PTblPatcher
 
                 foreach (var patch in CollectionsMarshal.AsSpan(patches))
                 {
-                    if (patch.SegmentDiffs.Count <= x) continue;
-                    var destination = GC.AllocateUninitializedArray<byte>(Math.Max(patch.SegmentDiffs[x].LengthAfterPatch, segments[x].Length));
-                    var patchDiff = patch.SegmentDiffs[x].Data;
-
-                    fixed (byte* destinationPtr = &destination[0])
-                    fixed (byte* currentSegmentPtr = segments[x].Span)
-                    fixed (byte* patchPtr = patchDiff.Span)
-                    {
-                        S56DiffDecoder.Decode(currentSegmentPtr, patchPtr, destinationPtr, (nuint)patch.SegmentDiffs[x].Data.Length, out var numWritten);
-                        segments[x] = new Memory<byte>(destination, 0, (int)numWritten);
-                    }
+                    if (patch.SegmentDiffs.Count <= x) 
+                        continue;
+                    
+                    ApplyPatch(patches, x, segments);
                 }
             }
 
@@ -218,35 +206,14 @@ public struct P3PTblPatcher
                 memoryStream.Write(segments[1].Length / 56); 
                 foreach(var segment in segments)
                     memoryStream.Write(segment.Span);
+                
                 return result;
             }
 
             foreach (var segment in segments)
-            {
-                memoryStream.Write(segment.Length);
-                memoryStream.Write(segment.Span);
-                memoryStream.AddPadding(P3PTblSegmentFinder.TblSegmentAlignment);
-            }
+                WriteSegment(memoryStream, segment, P3PTblSegmentFinder.TblSegmentAlignment);
 
             return result;
-        }
-    }
-
-    /// <summary>
-    /// Creates a diff for an individual segment and adds it to the patch.
-    /// </summary>
-    private unsafe void DiffSegment<T>(TblPatch patch, PointerLengthTuple newSegment, PointerLengthTuple originalSegment, T resolver) where T : IEncoderFieldResolver
-    {
-        var destination = GC.AllocateUninitializedArray<byte>((int)S56DiffEncoder.CalculateMaxDestinationLength(newSegment.Length));
-        fixed (byte* destinationPtr = destination)
-        {
-            var numEncoded = S56DiffEncoder.Encode(originalSegment.Pointer, newSegment.Pointer,
-                destinationPtr, (nuint)originalSegment.Length, (nuint)newSegment.Length, resolver);
-            patch.SegmentDiffs.Add(new TblPatch.SegmentDiff()
-            {
-                Data = destination.AsMemory(0, (int)numEncoded),
-                LengthAfterPatch = newSegment.Length
-            });
         }
     }
